requirements.txt
faiss-cpu
sentence-transformers
pypdf
langchain
langchain-ollama
langchain-text-splitters
ollama
jupyter
ipywidgets
